pcli: split up command definitions into submodules
This is the first step in refactoring the `pcli` code so that it's not all in
the `main` function.
